05_Line_Bot_theDays/requirements.txt
gunicorn
beautifulsoup4   == 4.9.0
Flask            == 1.1.2
googletrans      == 2.4.0
gTTS             == 2.1.1
pydub            == 0.23.1
SpeechRecognition == 3.8.1
line-bot-sdk     == 1.16.0
lxml             == 4.5.0
numpy            == 1.18.3
pandas           == 1.0.3
requests         == 2.23.0
PyMySQL		     == 0.9.3